// src/client/__generated__/types.ts
export interface GetAllUsers_getAllUsers {
  __typename: "User";
  id: string;
  email: string;
  name: string | null;
  role: UserRole;
}

export interface GetAllUsers {
  getAllUsers: GetAllUsers_getAllUsers[];
}
// ...
export interface UpdateUser_updateUser {
  __typename: "User";
  id: string;
  email: string;
  name: string | null;
  role: UserRole;
}

export interface UpdateUser {
  updateUser: UpdateUser_updateUser;
}

export interface UpdateUserVariables {
  id: string;
  input: UpdateUserInput;
}

/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: DeleteUser
// ====================================================

export interface DeleteUser_deleteUser {
  __typename: "User";
  id: string;
}

export interface DeleteUser {
  deleteUser: DeleteUser_deleteUser;
}

export interface DeleteUserVariables {
  id: string;
}
// ...
export enum UserRole {
  ADMINISTRATION = "ADMINISTRATION",
  EXECUTIVE = "EXECUTIVE",
  FACTORY = "FACTORY",
}
// ...
export interface UpdateUserInput {
  email: string;
  name?: string | null;
  password?: string | null;
  role?: UserRole | null;
}